Move storybook stories to live with their components for the office components thus far.
