+++ src/components/Office/allowancesTable.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, object } from '@storybook/addon-knobs';
 
-import AllowancesTable from '../components/Office/AllowancesTable';
+import AllowancesTable from './AllowancesTable';
 
 const info = {
   branch: 'Navy',
