+++ src/components/Office/ordersTable.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 
-import OrdersTable from '../components/Office/OrdersTable';
+import OrdersTable from './OrdersTable';
 
 storiesOf('TOO/TIO Components|OrdersTable', module)
   .addDecorator(withKnobs)
