+++ src/components/Office/shipmentContainer.stories.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 
-import ShipmentContainer from '../components/Office/ShipmentContainer';
-import ShipmentHeading from '../components/Office/ShipmentHeading';
+import ShipmentContainer from './ShipmentContainer';
+import ShipmentHeading from './ShipmentHeading';
 
 storiesOf('TOO/TIO Components|ShipmentContainer', module)
   .addDecorator(withKnobs)
