+++ src/components/Office/shipmentheading.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 
-import ShipmentHeading from '../components/Office/ShipmentHeading';
+import ShipmentHeading from './ShipmentHeading';
 
 storiesOf('TOO/TIO Components|ShipmentHeading', module)
   .addDecorator(withKnobs)
